[REF] l10n_es_reports: libros export handle unordered lines
Sometimes, when handling export of VAT Book (libros) by getting all the
lines from the tax report, some of the line might not be ordered as
expected. If a tax/surcharge line appear in the list before the base
line, this will result in a KeyError when trying to access
sheet_line_vals of the id because they have not been created yet.

This commit aims to fix that, and also refactor this part of the code to
be more clear, clean, & error-proof in the future. An handler for error
in the common error place (KeyError) is also written for similar
problems we might encounter in the future.

closes odoo/enterprise#56454

Task-id: 3703023
X-original-commit: abc47129d7dfb10c88c9e1a0d16cd52f137edc38
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>