[IMP] web_studio: batch studio approval fetching
Studio approval is the mechanism that allows to restrict some buttons
to some users until they are approved by a manager.

Before, the approval was fetched one by one, for each button in the
form view.

This commit changes this to fetch all the approvals in one request in
order to improve the performance.

The python code has been improved as well, making as few sql requests as possible.

The counterpart of this commit is a change of API of the public method
`studio.approval.rule:@get_approval_spec` that can handle fetching rules
for any combination of `model * res_ids * method * action_ids`

task-3601171

closes odoo/enterprise#58808

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>