[REM] *: remove useless date widget
Since odoo/odoo@b5794e89e1ad29e2a86c7dd in saas-16.3, the
widget date has no more effect on the rendering of a Datetime field.

For Date field, the only impact it has is in list views, where it uses
the component instead of formatting the value into a string. This only
requires more processing to reach the same result.

This commit therefore removes the use of this widget.

closes odoo/enterprise#55259

Related: odoo/odoo#151381
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>