[IMP] planning: update 'Allow Unassignment' option in Planning settings
Before this commit, it was not clear to users that the 'switch shifts'
feature existed, and that is it mutually exclusive with the 'allow unassignment'
option in the Planning settings.

To make things clear, this commit splits the 'allowing unassignment'
setting in two:
- The 'switch shifts with other employees' option which corresponds to the
current behavior when the 'allow unassignment' feature is disabled. It is
the default value.
- The 'unassignment themselves from shifts' option which corresponds to
the current behavior when the 'allow unassignment' feature is enabled.

task-3703648

closes odoo/enterprise#55260

Related: odoo/upgrade#5603
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>