[FIX] account_followup: email_from not set correctly in automatic followups
Before: the email_from was only used if the user manually executes the followup and was taken from the mail template's values in the send & print wizard.
Now: for automatic followups without a wizard, the email from is taken from the mail template on the followup line on the partner.
X-original-commit: a0e5112796be8374d8661bbe48d4a2bc94a96659
Part-of: odoo/enterprise#55954