[IMP] hr_appraisal: remove link to appraisal
we remove the link to appraisal from the form view, because
if you click on it, it will create a new appraisal, and this
goes against the original purpose of this warning that is to
try to not create duplicate appraisals. we also add a condition
for the warning to be displayed only if the record is not in the
database yet.

task-3582140

closes odoo/enterprise#55284

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>