[FIX] industry_fsm*: make tour work without demo-data
*: industry_fsm, industry_fsm_report

Before this commit, the python test launching the industry_fsm_tour was disabled
since version 16.4 because of a bug in the patch functions to add additional
steps in that tour but since 17.0, the issue is no longer there and so the test
could be executing again without demo data.

This commit removes the condition to skip the test once the demo data is not
installed since now the tour works as expected with and without demo data
installed.

closes odoo/enterprise#55358

X-original-commit: 6927d92d888c02b46590156c1ad93cee3487e5cc
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>