[FIX] web_grid: avoid altering {row,section,column} id in 2 rpc calls
Before this commit, when the load is slow and the user does not wait the
load method be executed and launches another search, the section, row and
column ids could be altered by the both load are currently executed
(the current one and the previous one whom is not yet finished). The problem
with that is the section, row and column id is a global variable in the file
and so the id is incremented by the both loads once a section, row or column
is created and so the user could see a big gap between the grid header and his
first result displayed because the row id could be for instance 21 instead of 1.

This commit fixes the issue by removing the global variable and define the
attributes to increment for section, row and column id into `GridDataPoint`
to be sure the id will only be incremented by only one load.

closes odoo/enterprise#55797

X-original-commit: fb91c45f5e4fc108982ae504c983144da0c6876a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>