[FIX] ups_delivery_rest: traceback on return label generation
This commit fixes some issues that prevented the user from
generating a return label in UPS REST module.

To reproduce:
Create a new shipping method with UPS. Enable "generate return
label" toggle, and attempt to validate a delivery.

There are a few issues as `ups_rest_get_return_label` was not
adapted when forward porting from 15 to 16:

+ `_prepare_shipping_data` is changed, so it does not return
`package_names`
+ `label_binary_data` from `_send_shipping` is changed from a
dictionary to a list
+ as `package_names` no longer is there, the log message needs
to be adapted.

This PR fixes these issues, so that the return label could be
generated.

opw-3689861

closes odoo/enterprise#55592

X-original-commit: b45fcce2b22c74e7c2c05d9382c92d95fcffa6f6
Signed-off-by: William Henrotin (whe) <whe@odoo.com>