[PERF] hr_referral: add title to referral link
this commit adds a 280x performance improvement on the
referral link generation.

Explaination:
The generation of title is quite costly for link.trackers because
it require to render the page to retrieve the metadata of it.
(more or less 1second per title) Since we don't actually care about
the title for hr_referral the url of the job is enough.

task-3707478

closes odoo/enterprise#56991

X-original-commit: ca8f0c4e99eddf43f39fee614a67ab79223b8eea
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>