[FIX] industry_fsm_sale: hide partner field in project sharing list view
Before this commit, the `hide_partner` context was passed by the python method
returning the main action for project sharing feature to be able to hide the
`partner_id` column in the main view when the project is not billable and nor
fsm one.

This commit overrides the list view to add that condition in `column_invisible`
attribute of `partner_id` field.

task-3615844

closes odoo/enterprise#55360

Related: odoo/odoo#147707
Signed-off-by: Audric Onockx (auon) <auon@odoo.com>