[FIX] website_appointment_sale: delete answers for bookings
Problem: When a user books an appointment, whose up-front payment is enabled,
and provides answers to the questions from the website,the user is unable
to remove the sale order line that contains the booking because a ValidationError
occurs due to the answers not getting deleted before the booking gets deleted.

Solution: When a sale order line or sale order that contains
an appointment booking gets deleted, its answers should be
deleted to prevent errors.

Steps to reproduce on Runbot:
1. Install Appointments and ecommerce
2. For the dental care appointment, edit and enable "Up-front payment"
3. On the website, book an appointment for dental care and proceed to payment
4. Attempt to remove the booking and a ValidationError occurs
"The operation cannot be completed: The answer inputs must be linked to a meeting or to a booking."

opw-3671511

closes odoo/enterprise#56614

X-original-commit: ef403581d27391d1138248aa4366b0dfeb98e99b
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>