[FIX] sign: high accuracy gps coordinates
The Location API allow developpers to specify options when requesting a
device's location - notably, it can request 'high accuracy' positioning.

By default, 'high accuracy' is set to false so that the location API
returns a position *quickly*. We can assume that it will return a
position based on e.g. WiFi networks it sees around it and an API call
to a location provider like Google, Apple, Here, etc. Setting 'high
accuracy' to high will make this process much slower (20-30s is not
gonna be an exception), but it means that devices with built-in GPS will
usually provide an far more precise location.

In Sign, it's rather important to have reliable data rather than a quick
process.

Assuming a person signing something for real, it's safe to assume that
the browser will have the time to provide a precise position before the
user submits their signature. If not, the position will simply not be
saved.

opw-3677499

closes odoo/enterprise#55656

X-original-commit: e6c7f405927a39ed072cf28c191566a77991a27f
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>