[PERF] spreadsheet_dashboard: avoid currency RPC
Most (if not all) dashboards have monetary amounts. They are formatted
with the main company currency format.
Before this commit, a RPC was made to fetch the company currency.
With this commit, the dashboard is loaded with the currency.

It saves one network request and a full spreadsheet evaluation (which would
have occured after the request is done)

closes odoo/enterprise#55415

Task: 3709466
Related: odoo/odoo#151725
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>