[FIX] room: use correct luxon locale right from loading
Purpose:
--------
Since the redesign of the frontend view (commit [1]), the locale used
when loading the frontend view is not the same as the one used once
the view is loaded.
This can be seen in the current datetime shown in the top right corner
(its format changes when it is updated the first time), or when a new
booking is added in the sidebar (it won't have the same time format as
the existing ones).

This commit ensures that the locale is correctly set when loading the
frontend view.

[1]: https://github.com/odoo/enterprise/commit/4e0cde83ed97d1f30d41d194b279ca62f2fea00a

Task-3609006

Part-of: odoo/enterprise#51351