[IMP] account_accountant: Reconcile wizard - partner reclassification
When the write-off's Account Type is either Receivable or Payable,
we allow the user to select a partner.
By default, the partner is the original one. The user can change it.
The write-off will then be on the original partner for the line that
balances the reconciled lines and the selected partner for the line
that goes on the selected account in the wizard.

task-3703287

closes odoo/enterprise#55518

Signed-off-by: Laurent Smet (las) <las@odoo.com>