[FIX] room: make frontend form view reactive to booking updates
Purpose:
--------
When a booking update notification is received while the user is creating
or editing a booking from the frontend form view, the edited booking's
duration is updated in the sidebar but not in the calendar view.

The view is not updated because it relies on the `onWillUpdateProps` hook
to recompute the `bookingsByDate` object used to render the view. However
when a booking is updated, the reference of the `bookings` object is not
changed, hence the `onWillUpdateProps` hook is not called.

This commit fixes this by adding a getter `bookingsByDate`so that the
bookings by date are recomputed on each rendering.

Task-3609006

Part-of: odoo/enterprise#51351