[FIX] l10n_es_reports: invoice incorrectly reported in mod349
Create an invoice for a customer who lives in the UE but his/her VAT number
has not been verified by the VIES system, so the transaction with this partner
should be considered with fiscal position 'Regimen Nacional' because the IVA
should appear as a domestic customer.
When creating the invoice with fiscal position 'Regimen Nacional', ensure
field for tax report 349 is empty in the tab 'AEAT data' as well
as tax in the invoice exists.
Open Tax Report 349

Issue: There will be values for these 2 lines which should not be there,
as this specific invoice should not appear in tax report 349

opw-3673456

closes odoo/enterprise#56218

X-original-commit: a9318df377d623d8395b8e69487321ff18eba560
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>