[FIX] hr_appraisal: Fix appraisal request url
To reproduce
============
- Go to Appraisal select an employee and send by email
- the url in the email is wrong

Problem
=======
- the body of the request is rendered twice with qweb, but the template
wasn't designed for that, as we have `t-att-href="ctx.get('url')"` that will
be `href="ctx['url']"` after first render and expecting to be rendered again
and replace `ctx['url']` with the real url but qweb has no way to know that
without a `t-att-` prefix.

Solution
========
- before second render, add `t-att-` prefix to `href` so qweb will know that
it should be rendered again and replace `ctx['url']` with the real url.

opw-3687928

closes odoo/enterprise#55812

X-original-commit: 5f2cd1c6c1048ff605d07e7f3af220bdccc05d2c
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>