[FIX] sign: traceback on document resend
After the conversion of the user service to a simple module, a traceback
started happening when trying to resend a document. This was due to the
missing conversion of the user service paradigm to the module one for that
specific flow. This commit fixes this by using the user module instead of the
user service.

task-3710633

closes odoo/enterprise#57029

X-original-commit: 053e723d3c087de88580cb34f940c908ebc633ce
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>