[FIX] sign: public document showing send button
After the sign conversion to OWL, when accessing an unsigned shared document it
would show a "send" button close to the public user. However, since it's a
shared document, there is no email to send the document to. Clicking on it,
will cause a traceback. This commit hides this button for shared requests.

task-3710635

closes odoo/enterprise#57170

X-original-commit: 400ad389591b709c354c220ba5c4cba23319780d
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>