[FIX] sign: public document showing send button
After the sign conversion to OWL, when accessing an unsigned shared document it
would show a "send" button close to the public user. However, since it's a
shared document, there is no email to send the document to. Clicking on it,
will cause a traceback. This commit hides this button for shared requests.

task-3710635

closes odoo/enterprise#57021

X-original-commit: 01f7b5cdd8706d7ac24961aa1da3089e8491e52c
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>