[FIX] room: do not show "stop" on midnight when selected
Purpose:
--------
Since the redesign of the frontend view (commit [1]), "start" and "stop"
hints are not shown in the start and end slots of the selected bookings
in the frontend form view anymore. It was still shown on the midnight
slot.
This commit removes it for the midnight slot too for consistency.

[1]: https://github.com/odoo/enterprise/commit/4e0cde83ed97d1f30d41d194b279ca62f2fea00a

Task-3609006

Part-of: odoo/enterprise#51351