[FIX] room: make view refresh instantly when current booking changes
Purpose:
--------
If the current booking is deleted, or if a user uses the "book now"
feature, the update is shown immediately in the sidebar but the view
is only updated on the next interval step making the app feel laggy,
and being not really convenient for future tests.

This commit ensures that the view is refreshed immediately in these
cases.

Task-3609006

Part-of: odoo/enterprise#51351