[FIX] knowledge: load Thread Component properly
This commit fixes a bug where the Comments would only load one of the different
Components present inside of the view, meaning that only one thread would show
their messages. This leads to possible losses of comments and the main chatter
being unable to show messages.

The issue was that both the comments and the form view shared a singular
chatter object in the environment. This object enables the Thread component to
know if it needs to either load more data or messages via two booleans:
`fetchData` and `fetchMessages` that would be set to false when the Thread
finished fetching either messages or data.

To fix this, the chatter in the environment below the Comment level was set to
false, as the condition to fetch data and messages is
`!this.env.chatter || this.env.chatter?.fetchData`.
This way when OWL mounts each Comment Component it will fetch the necessary
data without impacting each other and the Form view's main chatter.

task-3714345

closes odoo/enterprise#55517

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>