[FIX] web_map: add filter for partner for routing
When fetching the Routing in project->task, the map view shows the error "To get routing on
your map, you first need to set up your Mapbox token. -> Set up token" even if they already
have the token set up.

This happens when, for example:
 if you have 3 tasks where 2 of them have the customer field
filled and one doesn't. The function _fetchRoute tried to filter out the records that
don't have the latitude/longitude. But, here, since one of the records doesn't have a partner,
when it's trying to look into partner.partner_latitude, it throws an error. After the error,
mapBoxToken is set to '' in _partnerFetching, and called _openStreetMapAPI() resulting in
displaying the error, "to get routing on your map, you first need to set up your Mabox token."

So, adding record.partner in the filter will help filter out the records that don't
have partner.

To Reproduce on Runbot:
1. Set up mabox toekn
2. Go to Project->Task
3. Create 3 Tasks ( 2 with customer and 1 without customer )
4. Go to the map view and it'll disply the error.

Also, added a qunit test, and modified few lines in the fetchRoute function because the routes were being added based on the records without considering if the records have partner or not

closes odoo/enterprise#70618

X-original-commit: 601a5b6e881cd52260f68833d026b2b3814ed03d
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>