[FIX] l10n_lu_hr_payroll: Monthly declaration tests
These tests are testing the monthly declaration in 2022.
As of today, 2022 is no longer a valid year in the wizard, thus causing
the tests to fail.

In order to ensure that such an issue won't happen again, we will
freeze the time for these tests to dec 31 2022, so that the year will
always be valid.

closes odoo/enterprise#55620

X-original-commit: ddf2c0cef8ea2b5b6bd8a992ba8c6f20189513ed
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>