[IMP] l10n_au_hr_payroll: Ordinary Time Earnings
Implement rules for OTE, and update existing rules in order to
benefit from it and have a more correct value.

Also update the tests of the app in order to better test the expected payslips
for each structure, ensuring that future changes that will affect the payslips
values or line orders are made on purpose.

And finally, fix the issues that were brought to light by these new tests.

Task id # 3659302

closes odoo/enterprise#56993

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>