[REF] whatsapp: use `channel_type` rather than `type`
The goal is to remove the `type` field which corresponds to nothing in
database, and its value is fully deducible from other fields.

In particular for channels, `type` is always equal to `channel_type`.

Most of them are renamed in this commit, the remaining problematic ones
will be handled separately.

closes odoo/enterprise#55645

Related: odoo/odoo#152320
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>