[IMP] purchase, account, sale: add account.order(.line).mixin
A lot of behavior is shared between the sale order (line) and purchase
order (line). To avoid having to maintain this code in parallel and
allow for sharing of the overlapping code, we introduce two new mixins:
1. account.order.mixin, for sale.order and purchase.order
2. account.order.line.mixin, for sale.order.line and purchase.order.line

Additionally, this PR adds the option to create down payments for
purchase orders by moving the existing functionality from sale orders
into the newly created mixins. This new function is disabled by default
for purchase, but can be turned on with a new checkbox in the settings.

task-3600910

Part-of: odoo/enterprise#55711
Related: odoo/odoo#144608
Related: odoo/upgrade#5627
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>