[FIX] industry_fsm(_report): fix sign modal z-index
Steps:
Install industry_fsm, worksheet and website.
Create a task, set Joel Willis as assignee and add a worksheet.
Log in as Joel Willis, got to that task and click sign.

Issue:
Z-wise, the modal isn't above the header.

Cause:
The modal is at the same level as the sign button, that is in the
sidebar, which has `position: sticky;`. This fixes the `z-index` to the
default value (even setting it manually has no impact). The fact that
the modal's z-index is greater than the header's doesn't matter, as its
container limits it.

Solution:
Put the modal div in the main content, which will never be sticky.
Also, `mt-5 pt-5` were creating a "non grayed" stripe in the upper part
of the page (where the header is at), so we delete it.

task-3644729

closes odoo/enterprise#55630

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>