[FIX] account: reduce number of queries for reconciliation
Simply prefetch all the values because we are analyzing a list of
recordsets without any prefetch ids.

Also, reduce the time complexity of multiple pieces of code where a
filtered was done in a loop, leading to O(n^2) time complexity. (mainly
when there is an exchange difference)

closes odoo/enterprise#55633

Related: odoo/odoo#152284
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>