[FIX] account: reduce number of queries for reconciliation
Simply prefetch all the values because we are analyzing a list of
recordsets without any prefetch ids.

Also, reduce the time complexity of multiple pieces of code where a
filtered was done in a loop, leading to O(n^2) time complexity. (mainly
when there is an exchange difference)

closes odoo/enterprise#56348

X-original-commit: 59cc2f14244a6e0f79d046c0bb6c9c806c69782e
Related: odoo/odoo#153565
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>