[FIX] account_invoice_extract: fix incorrect boxes in validation
Previously, we were always returning a box in the validation request,
even when no box was clicked by the user.

Now, the box will only be part of the request if it was indeed clicked
by the user and that its text matches the text of the field.

This has little impact on the learning of the AI model as it was mostly
relying on the text of the validated value, not its box.

closes odoo/enterprise#55865

X-original-commit: fc380acc4e26fa7d599d404ae878af9b5f0e1462
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>