[FIX] hr_payroll: force unique id in the report
Steps to reproduce:
-------------------
- install hr_payroll and hr_holidays;
- create an employee;
- create a contract for this employee (since 1st January for example);
- create and approve a sick time off for this employee (for a day in January);
- create a payslip for this employee (the payslip has two "worked days");
- confirm the payslip;
- go to Reporting / Payroll and group by employee;
- click on the employee's line to display the list view.

Issue:
------
There are two records.
This can be explained by saying that this is a record by worked days type.
However, if we click on it, we see that the detail is the same for both records.

Cause:
------
The query which generates the virtual table `hr_payroll_report` will give as `id`,
the value of the `id` which corresponds to the payslip.
```sql
SELECT
	p.id as id,
	wd.id as wdid,
	wd.name
FROM
	(SELECT * FROM hr_payslip WHERE state IN ('done', 'paid')) p
	left join hr_payslip_worked_days wd on (wd.payslip_id = p.id)
```

The result of this query will be two records:
```
 id | wdid |    name
----+------+------------
  1 |   2  | Unpaid
  1 |   3  | Attendance
```

When we want to obtain the details of the record, we will perform a read
on the `hr.payroll.report` model for an `id` equal to 1 for both records,
i.e. we will retrieve the same values.

Solution:
---------
Forcing the id to be unique.
Add a field indicating the type to avoid confusion.

Note:
In the list view, it is possible to group by "Payslip Name" to avoid confusion.
Note 2:
It is a band-aid fix that helps us understand what is going on in the report,
but it needs to be redesigned to redirect us directly to payslip records
and not records corresponding to worked days (and avoid aggregation problems).

opw-3686692

closes odoo/enterprise#55950

X-original-commit: 099d9370deebfb159db4f57450a6ad466d91098e
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>