[FIX] social_twitter: show the full retweet message
This commit fixes an issue where the retweeted message would only display a
part of the retweeted message while twitter displays it in full.
This is a limitation of the Twitter API which only sends a limited part of the
tweet.
To fix this, we are setting the message value to the text of the referenced
tweet.

task-3653108

closes odoo/enterprise#59579

X-original-commit: 9f18b970ccdebfb095a170c111de82dfca8504d5
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>