[FIX] social_twitter: show the full retweet message
This commit fixes an issue where the retweeted message would only display a
part of the retweeted message while twitter displays it in full.
This is a limitation of the Twitter API which only sends a limited part of the
tweet.
To fix this, we are setting the message value to the text of the referenced
tweet.

task-3653108

closes odoo/enterprise#65171

X-original-commit: 1b0435a92492d85f28549a207fcdb7224b457c66
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>