[FIX] sale_planning: allocated_hours recomputed in onchange
Steps to reproduce:
-------------------
1. create a new database with industry_fsm_sale_report, planning and no
   demo data;
2. enable debug mode;
3. go to settings;
4. load demo data.

Issue:
------
Fails to load demo data due to a `KeyError` on
`planning.slot.allocated_hours`.

Cause:
------
When we update the module and instantiate the field, we will call
_recompute_all method but the method _compute_planning_hours_planned of
sale.order.line does an `add_to_compute` of planning.slot() field
allocated_hours.

So after _recompute_all, there is still something to compute. Then we
when we apply allocated_hours override (from sale_planning module) we
change the field object:

```py from models.py Model._setup_base
if len(fields_) == 1 and fields_[0]._direct and fields_[0].model_name == cls._name:
   cls._fields[name] = fields_[0] # we reuse the field
else:
   Field = type(fields_[-1])
   self._add_field(name, Field(_base_fields=fields_)) # new reference
```

and when we do the recompute, it is going to be done with the old field
reference that is now no longer on the model.

Solution:
---------
The issue comes from using add_to_compute inside a compute, in this
particular case (order of compute, field that is overridden, module
update that modify records/recompute values, ...) it is causing an
issue.

So to avoid this issue, we force the recomputation directly after the
add_to_compute.

Note:
-----
This is only happening in saas-16.4 and over because by chance, the
_recompute_all in saas-16.3 and below recomputes
sale.order.line.planning_hours_planned before
planning.slot.allocated_hours. So by chance, when we add allocated_hours
to be recomputed, it was already going to be done.

In saas-16.4 planning.slot.allocated_hours is computed before
planning_hours_planned and so the _recompute_all still has something to
compute after _recompute_all.

opw-3517083
opw-3691750

Closes odoo/odoo#139603

closes odoo/enterprise#59124

X-original-commit: 4b543aae52022cc52e7185b40f7e1a65d36e9163
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Co-authored-by: "Levi Siuzdak (sile)" <sile@odoo.com>