[FIX] account_asset: increase asset with constant period
When increasing an asset with constant periods, we get an error
on increase asset validation, due to the asset lifetime days
being wrongly computed.

Steps:

- Asset configuration:
	- Original value: 1200
	- Acquisition date: 10/01/2023
	- Method: linear
	- Duration: 12 Months
	- Computaion: Constant Periods
- Confirm asset
- Modify it: action 'modify' and date 01/15/2024
- Validate
-> UserError("The remaining value on the last depreciation line must be
0")

This is due to the fact that the field `asset_lifetime_days` is an
Integer field and it is computed to satisfy this constraint.
With this commit we change this field (non stored) from Integer
to Float and adapt its computation.
The tests already existing were not failing because the computation
of `asset_lifetime_days` always ended up on an integer as we mainly
tested increased assets on the 15th of months of 30 days.

opw-3691191

closes odoo/enterprise#55970

X-original-commit: 1eca8fc5a39260e149326811ddb9c6b601721c6c
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>