[IMP] account_reports, l10n_in: option/filter amount rounding
* Adds an option/filter to display amount in different units, i.e, 'decimals', 'units',
'thousands', 'lakhs' and 'millions'.

* Makes 'format_value' a private python method and creates a new public
'format_value' method that should be called from the client. This change was
made because we can not pass records from the client to the server.
For example for currency, we can pass the currency_id to the server, but then
we need to fetch the currency before continuing.

task-3626894

closes odoo/enterprise#57030

X-original-commit: 69e0d1600c22a07bdb2ea8dddf47db8bceaa743c
Related: odoo/odoo#154722
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>