[FIX] sale_subscription: double check `is_subscription` on write
Original commit prevented `subscription_state` from being true if the SO
wasn't a subscription. This fixed an issue with creating tasks/projects
for service products. As of 4ac62c700826ce1b38307641f91b7c9ce4f91d54,
keeping `subscription_state` as long as the quotation in draft status is
intended behavior for saas-16.4+. Instead the issue gets resolved on in
`action_confirm`.

This commit was amended to only add tests to verify this behavior.

opw-3709457

closes odoo/enterprise#56154

X-original-commit: eb23e24815b42359188c697e4d3e673c4fec5dfc
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>