[FIX] industry_fsm_stock: view product availability report
Before this commit, when the user is currently in the view to add some material
on a task and wants to view availability of a product, a traceback is occured
instead of displaying the report, the reason is because the context in the
action is a string and not a dict and so we cannot do an update to update the
context.

This commit fixes the issue by doing first a literal_eval to get a dictionary
for the action context and then update the context to add the warehouse to use.

closes odoo/enterprise#55747

X-original-commit: c78a49b0984fb36c9909cb02ca71ae4760912a79
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>