[FIX] sign: resend in mass tries to send shared requests
https://github.com/odoo/enterprise/pull/40112 allowed resending documents in
mass, however it does not check for shared sign requests. Since shared sign
requests are targetted at public users, it is not possible to resend them. This
PR fixes this by ignoring the shared sign requests. It was chosen to ignore
them because we still want to allow selecting all documents and resending only
the ones that can be resent.

task-3721338

closes odoo/enterprise#56982

X-original-commit: 6dd69eb932f5f8db8165ee4f29e76ee08526f1bd
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>