[FIX] whatsapp: improve variable display name computation
Before this commit, names of dynamic button variables are given `{{1}}`
statically. This name is appended to the `display_name` of button e.g.
`button 'button name' - {{1}}`. If there are two dynamic button variables
then both of the respective variables `display_name` will have `{{1}}` at
the end which doesn't make any sense.

After this commit names of dynamic button variables are the same as their
respective button name. So now the `display_name` is `Button - button name`
which is similar to location variables names and body variables names. Also
the `line_type` in the `display_name` of the variable is capitalized.

Task-3619004 (WhatsApp: Fix variable computation and name)

X-original-commit: odoo/enterprise@fad5a4d6550daebdba5d8381cd8082f34375dcd8
Part-of: odoo/enterprise#59987
Co-authored-by: Akshat Trivedi <aktr@odoo.com>