[FIX] voip: do not crash on cancelled call invitation
Sending a CANCEL request while a SIP session is being established will
result in a 487 Request Terminated response to the pending INVITE
request.

Prior to this commit, cancelling a SIP session would cause the session
to be deleted twice: once immediately after the CANCEL request was sent,
then a second time after the response to the INVITE request was
received, resulting in a crash the second time because the session had
already been deleted.

After this commit, 487 Request Terminated responses are ignored,
preventing the code from attempting to delete the already deleted
session and causing a crash.

opw-3702140

closes odoo/enterprise#55912

X-original-commit: 2efe0d24180e6e4afc2e95a95502b1eda859dba6
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>