[REF] spreadsheet_edition: move cell missing from spreadsheet
This commit moves the cell missing logic from the pivot model to
spreadsheet edition.

The objective is twofold:
* It's not necessary to recompute the missing cells when the spreadsheet
is not open in edition mode. (e.g. dashboard)

* The missing cells feature is a feature of spreadsheet_edition, so it
makes sense to move it to the spreadsheet_edition module.

This commit also changes the way the missing cells are computed. The
missing cells are now computed **only** when the user opens the
spreadsheet dialog to insert a specific pivot cell. This is more
efficient than recomputing the missing cells at each evaluation.

closes odoo/enterprise#55915

Task: 3724263
Related: odoo/odoo#152807
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>