[FIX] project_timesheet_forecast_sale: ensure the timesheet be validated
Before this commit, the timesheet to validate could be not found when
the filtered is made inside `action_validate_timesheet` method, it is
for that reason the timesheet created inside the test is not validated
as expected.

This commit ensures the timesheet created is recorded yesterday to be sure
the timesheet will be validated inside `action_validate_timesheet` method.

runbot-56444

closes odoo/enterprise#55924

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>