[FIX] l10n_au_hr_payroll: Fix demo data and tests
As seen in the nightly build https://runbot.odoo.com/runbot/build/57728091
There is an issue with the demo data where a hr employee record
is trying to be set in a res.partner field.

Also, one test is setting accounting stuff while this module
isn't dependent on any account modules.

closes odoo/enterprise#55932

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>