[FIX] hr_referral: add an uninstall hook
Issue:
------
When we uninstall the `hr_referral` module, an error occurs
when we want to create a job application.

This is because a domain which uses a field which only exists
if the `hr_referral` module is installed is persistent on window actions.

It is therefore necessary to remove these domains during uninstallation,
as they are no longer present on the `hr.applicant` model.

opw-3708569

closes odoo/enterprise#56113

X-original-commit: 6c33abc34434fa47d2f54fb09f285481b3412355
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>