[FIX] helpdesk: reload fieldsToCheck when record changed with pager
Before this commit, when the user disables a feature in helpdesk in all
helpdesk teams (by using the pager to switch helpdesk teams) the window is not
reloaded when it should be expected (when the SLA feature is disabled in all
helpdesk team). The reason is because the `fieldsToCheck` is not reset when the
record changed and so the variable keeps the value of the previous record.

This commit fixes the issue by adding a `onPatched` hook inside the custom
form controller to reset `fieldsToCheck`.
This commit also improves a bit the performance by only checking the changes
made by the user instead of checking the whole data of the record in the form
view to know if the view has to be reloaded or not.
Finally, it also calls the parent methods (`onWillRecordSaved` and
`onRecordSaved`) since those methods defined in `HelpdeskTeamFormController`
are methods overriden and not new methods in that class.

Issue found during the testing of task-3549328

closes odoo/enterprise#56325

X-original-commit: a954741bd2a3925a0316b138d4186f31d87fa9ea
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>